refactor and store data in mongodb
